Added a new enemy that shoots projectiles.
The enemy(spotter) can't be killed and contact with either the
projectile or his body do damage. Projectiles vanish after a set amount
of time, if they leave the screen up or down or if they hit a player or
a collidableTile.
Note: Not finished, contains bugs! (walking)

Added a base class for entities (though i already seem to be loading it
full of specific stuff...)